pyproject.toml
[project]
name = "telescope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black==24.4.2",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "isort==5.13.2",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation>=0.55b1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "ruff==0.4.10",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.3",
]
